refactor #5: 오타 수정
- 경유지 request DTo
- 지자체 멤버 번호 bigint -> smallint
- delete 응답 ok -> noContent

+++ src/main/java/com/cu2mber/stopoverservice/domain/Stopover.java
@@ -18,7 +18,7 @@ public class Stopover {
     @Column(name = "stopover_no")
     private Long stopoverNo;
 
-    @Column(name = "member_local_no", columnDefinition = "bigint", nullable = false)
+    @Column(name = "member_local_no", columnDefinition = "smallint", nullable = false)
     @Comment("지자체 멤버 번호")
     private Long memberLocalNo;
